+++ src/components/flatassistant/setBrightness.vue
@@ -3,8 +3,8 @@
     v-model="settingsStore.flats.brightness"
     :label="$t('components.flatassistant.brightness')"
     labelKey="components.flatassistant.brightness"
-    :min="0"
-    :max="100"
+    :min="store.flatdeviceInfo?.MinBrightness || 0"
+    :max="store.flatdeviceInfo?.MaxBrightness || 65535"
     :step="1"
     :decimalPlaces="0"
     inputId="brightness"
@@ -13,6 +13,8 @@
 <script setup>
 import { useSettingsStore } from '@/store/settingsStore';
 import NumberInputPicker from '@/components/helpers/NumberInputPicker.vue';
+import { apiStore } from '@/store/store';
+const store = apiStore();
 
 const settingsStore = useSettingsStore();
 </script>